Refactor and add rpm builds
This change refactors shared functions into a single lib/ directory. The
dockerfiles will now modify PYTHONPATH to properly include the files in
this directory. This allows common functions, like github operations
and running cli commands, to be shared across tests and builds.
